feat: Added Authorized request to add,delete or fetch Data

+++ frontend/src/components/WorkoutDetails.js
@@ -1,14 +1,24 @@
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext";
+import { useAuthContext } from "../hooks/useAuthContext";
 
 // Date fns
 import { formatDistanceToNow } from "date-fns";
 
 const WorkoutDetails = ({ workout }) => {
   const { dispatch } = useWorkoutsContext();
+  const { user } = useAuthContext();
 
   const handleClick = async () => {
+    // Check if user is logged in if not exit
+    if (!user) {
+      return;
+    }
+
     const response = await fetch(`/api/workouts/${workout._id}`, {
       method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${user.token}`, // Add the token to the header, it's a Bearer token
+      },
     });
     const json = await response.json();
